refactor: reorganize test files and folders
