+++ backend/src/adapters/http/status_activity.py
@@ -3,7 +3,7 @@ from typing import Annotated
 from uuid import UUID
 
 from fastapi import APIRouter, Depends, Header, HTTPException, Path, status
-from src.adapters.input.http.schemas import ActivityResponse, StatusResponse
+from src.adapters.http.schemas import ActivityResponse, StatusResponse
 from src.application.ports.input.get_activity_status_port import GetActivityStatusPort
 from src.config.dependencies import get_status_activity_service
