+++ backend/src/adapters/persistence/activity_document.py
