+++ backend/src/adapters/websocket/activity_stream.py
@@ -6,7 +6,7 @@ from uuid import UUID
 from fastapi import APIRouter, Depends, Path, Query, WebSocket, WebSocketDisconnect
 from fastapi.websockets import WebSocketState
 from pydantic.json import pydantic_encoder
-from src.adapters.input.websocket.schemas import ActivityStream, ActivityStreamType
+from src.adapters.websocket.schemas import ActivityStream, ActivityStreamType
 from src.application.ports.input.get_activity_status_port import GetActivityStatusPort
 from src.config.dependencies import get_status_activity_service
 from src.domain.entities.activity import Activity
@@ -76,6 +76,7 @@ async def _build_payload(
 ) -> str:
     try:
         result = await service.get_activity_status(activity_id, participant_id)
+    
         if isinstance(result, Activity):
             message = {
                 "type": (
