+++ backend/src/adapters/websocket/router_ws.py
@@ -1,0 +1,6 @@
+from fastapi import APIRouter
+from src.adapters.websocket.activity_stream import router as ws_activity_status_router
+
+ws_router = APIRouter(prefix="/v1")
+
+ws_router.include_router(ws_activity_status_router)
