+++ backend/src/main.py
@@ -3,7 +3,8 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from src.adapters.input.router import router
+from src.adapters.http.router_http import http_router
+from src.adapters.websocket.router_ws import ws_router
 from src.config.database import initialize_database
 from src.config.logger import initialize_logger
 from src.config.monitoring import initialize_monitoring
@@ -37,7 +38,8 @@ def create_app() -> FastAPI:
         allow_headers=["*"],
     )
 
-    app.include_router(router)
+    app.include_router(http_router)
+    app.include_router(ws_router)
 
     return app
