+++ backend/tests/integration/http/test_result_activity_route.py
@@ -1,3 +1,4 @@
+from http import HTTPStatus
 from unittest.mock import patch
 
 import pytest
@@ -7,44 +8,52 @@ import pytest
 @pytest.fixture
 def mock_repository_find_one():
     with patch(
-        "src.adapters.output.activity_repository_mongo_adapter.ActivityRepositoryAdapter.find_one"
+        "src.adapters.persistence.activity_repository_mongo_adapter.ActivityRepositoryAdapter.find_one"
     ) as mock_find_one:
         yield mock_find_one
 
 # --- Tests ---
 
 @pytest.mark.asyncio
-async def test_status_activity_success(
+async def test_result_activity_success(
     async_client,
     mock_activity_fixture,
     mock_activity_document_fixture,
-    mock_participant_regular,
-    mock_participant_owner,
 ):
     # Given
     await mock_activity_document_fixture.insert()
 
     # When
     response = await async_client.get(
-        f"/v1/activity/{mock_activity_fixture.id}",
-        headers={"X-Participant-Id": str(mock_participant_regular.id)}
+        f"/v1/activity/{mock_activity_fixture.id}/result",
     )
 
     # Assert
-    assert response.status_code == 200
-    assert response.json()["activity"]["owner"]["id"] == str(mock_participant_owner.id)
+    assert response.status_code == HTTPStatus.OK
+
+    body = response.json()
+    assert body["activity"]["activity_id"] == str(mock_activity_fixture.id)
+
+    result = body.get("result")
+    assert result is not None, "'result' must be present in the response"
+
+    mandatory_result_keys = {"overall_score", "dimension_scores"}
+    assert mandatory_result_keys.issubset(result), "missing keys in 'result'"
+
+    dim_scores = result.get("dimension_scores")
+    assert isinstance(dim_scores, list) and dim_scores, (
+        "'dimension_scores' must be a non-empty list"
+    )
 
 
 @pytest.mark.asyncio
-async def test_status_activity_not_found(
+async def test_result_activity_not_found(
     async_client,
-    mock_activity_fixture,
-    mock_participant_regular,
+    mock_activity_fixture
 ):
     # When
     response = await async_client.get(
-        f"/v1/activity/{mock_activity_fixture.id}",
-        headers={"X-Participant-Id": str(mock_participant_regular.id)}
+        f"/v1/activity/{mock_activity_fixture.id}/result",
     )
 
     # Assert
@@ -56,7 +65,6 @@ async def test_status_activity_not_found(
 async def test_status_activity_unexpected_error(
     async_client,
     mock_activity_fixture,
-    mock_participant_regular,
     mock_repository_find_one
 ):
     # Given
@@ -64,8 +72,7 @@ async def test_status_activity_unexpected_error(
 
     # When
     response = await async_client.get(
-        f"/v1/activity/{mock_activity_fixture.id}",
-        headers={"X-Participant-Id": str(mock_participant_regular.id)}
+        f"/v1/activity/{mock_activity_fixture.id}/result",
     )
 
     # Assert
