+++ backend/tests/unit/adapters/persistence/test_activity_document.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING
 
 import pytest
-from src.adapters.output.activity_document import ActivityDocument
+from src.adapters.persistence.activity_document import ActivityDocument
 
 if TYPE_CHECKING:
     from src.domain.entities.activity import Activity
