+++ backend/tests/unit/adapters/persistence/test_activity_repository_mongo_adapter.py
@@ -2,8 +2,8 @@ from dataclasses import replace
 from unittest.mock import AsyncMock, patch
 
 import pytest
-from src.adapters.output.activity_document import ActivityDocumentForMongo
-from src.adapters.output.activity_repository_mongo_adapter import ActivityRepositoryAdapter
+from src.adapters.persistence.activity_document import ActivityDocumentForMongo
+from src.adapters.persistence.activity_repository_mongo_adapter import ActivityRepositoryAdapter
 
 
 @pytest.fixture
